requirements.txt
dnspython>=2.3.0
requests>=2.31.0
python-whois>=0.8.0
selenium>=4.15.0